stats: port Stats_AddDistanceTravelled to TRX
