+++ src/tr2/game/stats.c
@@ -217,12 +217,3 @@ void Stats_AddMedipacksUsed(const double medipack_value)
         Savegame_GetCurrentInfo(Game_GetCurrentLevel());
     current_info->stats.medipacks_used += medipack_value;
 }
-
-void Stats_AddDistanceTravelled(const XYZ_32 pos, const XYZ_32 last_pos)
-{
-    RESUME_INFO *const current_info =
-        Savegame_GetCurrentInfo(Game_GetCurrentLevel());
-    current_info->stats.distance_travelled += Math_Sqrt(
-        SQUARE(pos.z - last_pos.z) + SQUARE(pos.y - last_pos.y)
-        + SQUARE(pos.x - last_pos.x));
-}
