+++ src/tr2/game/stats.h
@@ -20,4 +20,3 @@ void Stats_AddKill(void);
 void Stats_AddAmmoHits(void);
 void Stats_AddAmmoUsed(void);
 void Stats_AddMedipacksUsed(double medipack_value);
-void Stats_AddDistanceTravelled(XYZ_32 pos, XYZ_32 last_pos);
